Started adding option to save results into subfolder of results. Currently only supports DateTime method but was written so that other subfolder naming methods could be included in the future.
Updated json format constants and created new json file. Included case to check if the subfolder variable exists within analysis_params, to then check if it is set to true, if true checks if the method is 'DateTime', then creates a purepath with the current time stamp in YYYY_MM_DD_HH_MM_SS format.

Next step actually make the dir and save to it later within the code.